+++ R/MUD.R
@@ -1,18 +1,26 @@
 #'
-#'
 #' Made Up 2D NMR-Like Data Sets
 #' 
-#' Made Up Data that resemble a simple, HSQC-like 2D NMR data set.  Lean, low resolution 
+#' Made Up Data that resemble simple, HSQC-like 2D NMR data sets.  Lean, low resolution 
 #' and designed primarily to check graphics and test functions. \pkg{As this is made up
 #' data, there is no underlying tri-linear structure and therefore one should NOT try to interpret
 #' the output of \code{miaSpectra2D} or \code{pfacSpectra2D} run on this data}.
-#' Real 2D NMR data sets are too large to be included in the package.
-#' 
-#' \code{MUD1} is composed of 10 samples in two groups.
-#' 
+#' \itemize{
+#'   \item \code{MUD1} is intended to test and demonstrate data reduction functions.  The HSQC-like data
+#' is derived from the 1H and 13C spectra of 3-methyl-1-butanol and the corresponding ethyl ether,
+#' idealized slightly for simplicity. There are 10 spectra.  Sample 1 is the alcohol; samples 2-5 are
+#' the alcohol with local shifts (specifically, two peaks have been shifted +/- one data point).
+#' Samples 6-10 are the ether, treated in a similar fashion. 
+#'   \item \code{MUD2} is intended to test and demonstrate alignment algorithms.  The HSQC-like data
+#' is derived from the 1H and 13C spectra of 3-methyl-1-butanol, idealized slightly for
+#' simplicity. There are 10 spectra.  The first one is "correct" and the other samples have global shifts on one
+#' or both dimensions.
+#' }
+#'
 #' @author Bryan A. Hanson, DePauw University.
 #'
-#' @name MUD1
+#' @name MUD
+#' @aliases MUD1 MUD2
 #'
 #' @docType data
 #'
@@ -22,10 +30,7 @@
 #'
 #' @keywords datasets
 #'
-#' @examples
-#'
-#' data(MUD1)
-#' sumSpectra(MUD1)
+#' @seealso These data sets are used in the examples of many functions.
 #'
 NULL