+++ R/inspectLvls.R
@@ -6,7 +6,7 @@
 #' used in the histogram so you can choose levels that are suitable for any
 #' particular spectrum.  However, loading matrices are excluded unless specified 
 #' as the input. Any of the arguments to \code{\link{calcLvls}} can be used
-#' to capture the computed levels, or you can choose by inspection.
+#' to compute the levels, or you can choose by inspection.
 #' 
 #' @param spectra An object of S3 class \code{\link{Spectra2D}}.
 #'
@@ -28,8 +28,8 @@
 #' @examples
 #' 
 #' data(MUD1)
-#' inspectLvls(MUD1, ylim = c(0, 50), main = "All MUD1 Data, mode = even")
-#' inspectLvls(MUD1, ylim = c(0, 50), mode = "NMR",  main = "All MUD1 Data, mode = NMR")
+#' inspectLvls(MUD1, ylim = c(0, 300), main = "All MUD1 Data, mode = even")
+#' inspectLvls(MUD1, ylim = c(0, 300), mode = "NMR",  main = "All MUD1 Data, mode = NMR")
 #' 
 inspectLvls <- function(spectra, loading = NULL, ...) {
 
@@ -45,8 +45,8 @@ inspectLvls <- function(spectra, loading = NULL, ...) {
   if (!is.null(loading)) {
   	pat <- paste("Loading", loading, sep = "_")
   	Ldone <- grep(pat, spectra$names)
-  	if (length(Ldone) == 0) stop("This Spectra2D object did not contain loadings.\n\t
-  	  You need to run miaLoadings or pfacLoadings first.")
+  	if (length(Ldone) == 0) stop("This Spectra2D object did not contain loadings.
+  	  You need to run plotLoadings2D and save the results.")
   	lvls <- calcLvls(spectra$data[[Ldone]], showHist = TRUE, ...)
    }